Código refatorado para a nova estrutura de pastas
-- Refatorado código para a nova estrutura de pastas, sendo o root direto em src ao invés de src/server
-- Deletado o package 'client' e todos os seus conteúdos
