--- src/client/CLI.java
@@ -1,5 +1,0 @@
-public class CLI {
-	public static void main(String args[]) {
-		System.out.println("Hello World!\nThis is the CLI main class");
-	}
-}
